+++ lib/dartapi_auth.dart
@@ -2,3 +2,5 @@ library;
 
 export 'src/jwt_service.dart';
 export 'src/auth_middleware.dart';
+export 'src/token_store.dart';
+export 'src/api_key_middleware.dart';